+++ persistent/__init__.py
@@ -28,38 +28,21 @@ __all__ = [
     'PickleCache',
     'TimeStamp',
 ]
-from persistent._compat import PURE_PYTHON
+
 from persistent.interfaces import IPersistent
-from persistent.interfaces import IPickleCache
 
 import persistent.timestamp as TimeStamp
 
-from persistent import persistence as pyPersistence
-from persistent import picklecache as pyPickleCache
-
-try:
-    # Be careful not to shadow the modules
-    from persistent import cPersistence as _cPersistence
-    from persistent import cPickleCache as _cPickleCache
-except ImportError: # pragma: no cover
-    _cPersistence = None
-    _cPickleCache = None
-else:
-    # Make an interface declaration for Persistent
-    # Note that the Python version already does this.
-    from zope.interface import classImplements
-    classImplements(_cPersistence.Persistent, IPersistent)
-    classImplements(_cPickleCache.PickleCache, IPickleCache)
-
-
-_persistence = pyPersistence if PURE_PYTHON or _cPersistence is None else _cPersistence
-_picklecache = pyPickleCache if PURE_PYTHON or _cPickleCache is None else _cPickleCache
+from persistent import persistence as _persistence
+from persistent import picklecache as _picklecache
 
 Persistent = _persistence.Persistent
+PersistentPy = _persistence.PersistentPy
 GHOST = _persistence.GHOST
 UPTODATE = _persistence.UPTODATE
 CHANGED = _persistence.CHANGED
 STICKY = _persistence.STICKY
 PickleCache = _picklecache.PickleCache
+PickleCachePy = _picklecache.PickleCachePy
 
 sys.modules['persistent.TimeStamp'] = sys.modules['persistent.timestamp']